+++ subot/settings.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from enum import Enum, auto
 from pathlib import Path
 
 from sqlalchemy import create_engine
@@ -19,3 +20,29 @@ IMAGE_PATH = Path(__file__).parent.parent.joinpath('resources')
 DEBUG = False
 VIEWER = False
 FPS = 20
+
+
+
+class GameControl(Enum):
+    CANCEL = auto()
+    CONFIRM = auto()
+    DOWN = auto()
+    LEFT = auto()
+    OPTION = auto()
+    RIGHT = auto()
+    UP = auto()
+
+
+keyboard_controls = {
+    'w': GameControl.UP,
+    'up': GameControl.UP,
+    's': GameControl.DOWN,
+    'down': GameControl.DOWN,
+    'd': GameControl.RIGHT,
+    'right': GameControl.RIGHT,
+    'a': GameControl.LEFT,
+    'left': GameControl.LEFT,
+    'e': GameControl.CONFIRM,
+    'q': GameControl.CANCEL,
+    'f': GameControl.OPTION,
+}